+++ src/rx-database.ts
@@ -37,7 +37,6 @@ import {
     PROMISE_RESOLVE_FALSE,
     randomCouchString,
     ensureNotFalsy,
-    PROMISE_RESOLVE_VOID,
     getDefaultRevision,
     getDefaultRxDocumentMeta,
     defaultHashFunction
@@ -71,16 +70,14 @@ import {
     INTERNAL_STORAGE_NAME
 } from './rx-storage-helper';
 import type { RxBackupState } from './plugins/backup';
-import {
-    createRxCollectionStorageInstance
-} from './rx-collection-helper';
 import { ObliviousSet } from 'oblivious-set';
 import {
     ensureStorageTokenDocumentExists,
     getAllCollectionDocuments,
     getPrimaryKeyOfInternalDocument,
     INTERNAL_CONTEXT_COLLECTION,
-    INTERNAL_STORE_SCHEMA
+    INTERNAL_STORE_SCHEMA,
+    _collectionNamePrimary
 } from './rx-database-internal-store';
 
 /**
@@ -293,7 +290,8 @@ export class RxDatabaseBase<
                     name: collectionName as any,
                     schemaHash: schema.hash,
                     schema: schema.jsonSchema,
-                    version: schema.version
+                    version: schema.version,
+                    connectedStorages: []
                 },
                 _deleted: false,
                 _meta: getDefaultRxDocumentMeta(),
@@ -366,52 +364,6 @@ export class RxDatabaseBase<
         );
 
         return ret;
-    }
-
-    /**
-     * delete all data of the collection and its previous versions
-     */
-    removeCollection(collectionName: string): Promise<void> {
-        let destroyPromise = PROMISE_RESOLVE_VOID;
-        if ((this.collections as any)[collectionName]) {
-            destroyPromise = (this.collections as any)[collectionName].destroy();
-        }
-        // remove schemas from internal db
-        return destroyPromise
-            .then(() => _removeAllOfCollection(this as any, collectionName))
-            // get all relevant pouchdb-instances
-            .then(knownVersions => {
-                return Promise.all(
-                    knownVersions
-                        .map(knownVersionDoc => {
-                            return createRxCollectionStorageInstance(
-                                this.asRxDatabase,
-                                {
-                                    databaseInstanceToken: this.token,
-                                    databaseName: this.name,
-                                    collectionName,
-                                    schema: knownVersionDoc.data.schema,
-                                    options: this.instanceCreationOptions,
-                                    multiInstance: this.multiInstance
-                                }
-                            );
-                        })
-                );
-            })
-            // remove the storage instance
-            .then(storageInstances => {
-                return Promise.all(
-                    storageInstances.map(
-                        instance => instance.remove()
-                    )
-                );
-            })
-            .then(() => runAsyncPluginHooks('postRemoveRxCollection', {
-                storage: this.storage,
-                databaseName: this.name,
-                collectionName
-            }))
-            .then(() => { });
     }
 
     /**
@@ -556,44 +508,6 @@ function throwIfDatabaseNameUsed(
 }
 
 /**
- * returns the primary for a given collection-data
- * used in the internal pouchdb-instances
- */
-export function _collectionNamePrimary(name: string, schema: RxJsonSchema<any>) {
-    return name + '-' + schema.version;
-}
-
-/**
- * removes all internal docs of a given collection
- * @return resolves all known collection-versions
- */
-export async function _removeAllOfCollection(
-    rxDatabase: RxDatabaseBase<any, any, any>,
-    collectionName: string
-): Promise<RxDocumentData<InternalStoreCollectionDocType>[]> {
-    const docs = await getAllCollectionDocuments(
-        rxDatabase.storage,
-        rxDatabase.internalStore
-    );
-    const relevantDocs = docs
-        .filter((colDoc) => colDoc.data.name === collectionName);
-    const writeRows = relevantDocs.map(doc => {
-        const writeDoc = flatCloneDocWithMeta(doc);
-        writeDoc._deleted = true;
-        return {
-            previous: doc,
-            document: writeDoc
-        };
-    });
-    return rxDatabase.internalStore
-        .bulkWrite(
-            writeRows,
-            'rx-database-remove-collection-all'
-        )
-        .then(() => relevantDocs);
-}
-
-/**
  * Creates the storage instances that are used internally in the database
  * to store schemas and other configuration stuff.
  */
@@ -732,7 +646,7 @@ export async function removeRxDatabase(
     );
 
     const collectionDocs = await getAllCollectionDocuments(
-        storage,
+        storage.statics,
         dbInternalsStorageInstance
     );